Rakefile modified to compile dict Ahead-of-Time (AOT) with macruby

+++ Rakefile
@@ -13,9 +13,22 @@ end
 
 task :install do
   sudo
+  
+  if File.exist? '/usr/local/bin/macrubyc' # this should probably be the rake default task.
+    Dir.mkdir './.tmp/'
+    File.open './.tmp/dict.rb', 'w' do |f|
+      f.print File.read './bin/dict'
+    end
+    `/usr/local/bin/macrubyc -o dict ./.tmp/dict.rb`
+    `rm -rf ./.tmp/`
+    `install -b ./dict /usr/local/bin`
+    File.delete './dict'
+  end
     
   Dir.mkdir '/usr/local/bin/' if not File.exist? '/usr/local/bin/'
-  system 'install -b '+Dir.nonhidden_entries('./bin/').map {|x| './bin/'+x}.join(' ')+' /usr/local/bin/'
+  bins = Dir.nonhidden_entries('./bin/')
+  bins.delete 'dict'
+  system 'install -b '+bins.map {|x| './bin/'+x}.join(' ')+' /usr/local/bin/'
 end
 
 task :uninstall do
@@ -23,8 +36,10 @@ task :uninstall do
   
   Dir.nonhidden_entries('./bin/').each do |x|
     f = '/usr/local/bin/'+x
-    puts "Removing #{f} ..."
-    File.delete f
+    if File.exist? f
+      puts "Removing #{f} ..."
+      File.delete f
+    end
   end
 end
